requirements.txt
# Trading Strategy Framework Dependencies
# Core data processing and analysis
pandas>=1.5.0
numpy>=1.23.0

# Configuration and data formats
pyyaml>=6.0
python-dateutil>=2.8.0

# Visualization and plotting
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.0.0

# Statistical analysis and scientific computing
scipy>=1.9.0
scikit-learn>=1.1.0

# Performance and optimization
numba>=0.56.0
joblib>=1.2.0

# Data validation and schemas
jsonschema>=4.0.0

# Networking and web requests (for data fetching)
requests>=2.28.0
aiohttp>=3.8.0

# Crypto/trading specific (optional but commonly needed)
ccxt>=3.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Jupyter notebook support (for evaluator agent)
jupyter>=1.0.0
ipykernel>=6.0.0

# Progress bars and utilities
tqdm>=4.64.0

# Memory profiling and performance monitoring
psutil>=5.9.0
memory-profiler>=0.60.0